add search geopos + error checks

+++ src/services/address.service.ts
@@ -9,7 +9,10 @@ const addressService = {
 	},
 
 	findByUserID: async (id: string) => {
-        const idUser = parseInt(id); 
+        const idUser = parseInt(id);
+        if (!isNaN(idUser)){
+            throw new Api400Error(`invalid Mission id ( = \'${id}\')`);
+        }
         return AddressRepository.find({
             where : { 
                 id_user: idUser
@@ -18,6 +21,9 @@ const addressService = {
     },
     findByMissionID: async (id: string) => {
         const idMission = parseInt(id); 
+        if (!isNaN(idMission)){
+            throw new Api400Error(`invalid Mission id ( = \'${id}\')`);
+        }
         return AddressRepository.find({
             where : { 
                 id_mission: idMission
@@ -27,15 +33,21 @@ const addressService = {
 	create : async (address : Address) => {
 		return AddressRepository.save(address);
 	},
-    update : async (address : Address) => {  // save fait les 2 donc bon ..
-		return AddressRepository.save(address);
+    update : async (address : Address) => {
+        const res = await AddressRepository.save(address);
+        if (res.id != address.id){
+            await AddressRepository.delete(res.id);
+            throw new Api400Error(`No existing Address for this id ( = \'${address.id}\')`);
+        }
+        return res;
 	},
     delete : async ( id: string) => {
         const idAddress = parseInt(id);
+        if (!isNaN(idAddress)){
+            throw new Api400Error(`invalid Address id ( = \'${id}\')`);
+        }
         return AddressRepository.delete(idAddress);
     }
-
-
 }
 
 module.exports = addressService;
